Authentication support using OAuth2 has been implemented. The applica…

+++ app/src/main/java/com/mosz/goposcodingtask/model/submodels/Tax.kt
@@ -4,5 +4,5 @@ data class Tax(
     val code: String,
     val id: Int,
     val name: String,
-    val rate: Int
+    val rate: Double
 )
